third_party/WebKit/LayoutTests/fast/xpath/detached-subtree-invalidate-iterator-expected.txt

Test that iterators are invalidated even if the original context is detached.

PASS result.invalidIteratorState is false
PASS result.iterateNext().tagName is 'span'
Modifying the document...
PASS result.invalidIteratorState is true
PASS result.iterateNext() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS successfullyParsed is true

TEST COMPLETE
